+++ packages/client/src/phaser/scenes/Preloader.ts
@@ -54,6 +54,6 @@ export class Preloader extends Scene {
   }
 
   create() {
-    this.scene.start('Game')
+    this.scene.start('Main')
   }
 }
